requirements.txt
git+https://github.com/Rapptz/discord.py
discord.py[voice]
PyNaCl==1.4.0
dnspython==2.0.0
async-timeout==3.0.1
youtube_dl
youtube-search-python
python-dotenv